+++ content_service/src/main/java/io/urdego/content_service/ContentServiceApplication.java
@@ -2,7 +2,9 @@ package io.urdego.content_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class ContentServiceApplication {
