--- content_service/src/main/java/io/urdego/content_service/common/config/SwaggerConfig.java
@@ -1,20 +1,0 @@
-package io.urdego.content_service.common.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(
-                        new Info()
-                                .title("Group Service API")
-                                .description("Documentation for Group Service API")
-                                .version("v1.0"));
-    }
-}
