+++ content_service/src/main/java/io/urdego/content_service/common/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package urdego.common.exception;
+package io.urdego.content_service.common.exception;
 
 public record ErrorResponse(int status, String title, String message) {
     public static ErrorResponse from(int status, String title, String message) {
