+++ content_service/src/main/java/io/urdego/content_service/common/exception/aws/AwsException.java
@@ -1,0 +1,10 @@
+package io.urdego.content_service.common.exception.aws;
+
+import io.urdego.content_service.common.exception.ContentException;
+import io.urdego.content_service.common.exception.ExceptionMessage;
+
+public class AwsException extends ContentException {
+    public AwsException(ExceptionMessage message) {
+        super(message.getText());
+    }
+}
