+++ content_service/src/main/java/io/urdego/content_service/domain/entity/user/constant/ContentInfo.java
@@ -12,23 +12,23 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ContentInfo {
 
-    @Column(name = "place_name")
-    private String placeName;
+    @Column(name = "content_type")
+    private String contentType; // 파일 콘텐츠 타입 (ex: image/jpg)
 
-    @Column(name = "height")
-    private int height; // 높이
+    @Column(name = "meta_latitude")
+    private Double metaLatitude; // 컨텐츠 위치정보
 
-    @Column(name = "width")
-    private int width; // 너비
+    @Column(name = "meta_longitude")
+    private Double metaLongitude;
 
     @Column(name = "size")
     private long size; // 사진 크기
 
     @Builder
-    public ContentInfo(String placeName, int height, int width, long size) {
-        this.placeName = placeName;
-        this.height = height;
-        this.width = width;
+    public ContentInfo(String contentType, Double metaLatitude, Double metaLongitude, long size) {
+        this.contentType = contentType;
+        this.metaLatitude = metaLatitude;
+        this.metaLongitude = metaLongitude;
         this.size = size;
     }
 }
