+++ content_service/src/main/java/io/urdego/content_service/domain/entity/user/repository/UserContentRepository.java
@@ -1,7 +1,7 @@
 package io.urdego.content_service.domain.entity.user.repository;
 
-import io.urdego.content_service.domain.entity.user.UserContent;
-
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import io.urdego.content_service.domain.entity.user.UserContent;
 
 public interface UserContentRepository extends JpaRepository<UserContent, Long> {}
