--- content_service/src/main/java/urdego/ContentServiceApplication.java
@@ -1,14 +1,0 @@
-package urdego;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@SpringBootApplication
-public class ContentServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ContentServiceApplication.class, args);
-    }
-}
