--- content_service/src/main/java/urdego/common/exception/ContentException.java
@@ -1,7 +1,0 @@
-package urdego.common.exception;
-
-public abstract class ContentException extends RuntimeException {
-    public ContentException(String message) {
-        super(message);
-    }
-}
