--- content_service/src/main/java/urdego/common/exception/aws/AwsException.java
@@ -1,10 +1,0 @@
-package urdego.common.exception.aws;
-
-import urdego.common.exception.ContentException;
-import urdego.common.exception.ExceptionMessage;
-
-public class AwsException extends ContentException {
-    public AwsException(ExceptionMessage message) {
-        super(message.getText());
-    }
-}
