--- content_service/src/main/java/urdego/common/exception/user/UserContentException.java
@@ -1,9 +1,0 @@
-package urdego.common.exception.user;
-
-import urdego.common.exception.ContentException;
-
-public class UserContentException extends ContentException {
-    public UserContentException(String message) {
-        super(message);
-    }
-}
