--- content_service/src/main/java/urdego/domain/entity/BaseEntity.java
@@ -1,27 +1,0 @@
-package urdego.domain.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-
-import lombok.Getter;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public abstract class BaseEntity {
-
-    @CreatedDate
-    @Column(name = "created_date_time")
-    private LocalDateTime createdDateTime;
-
-    @LastModifiedDate
-    @Column(name = "modified_date_time")
-    private LocalDateTime modifiedDateTime;
-}
